+++ lua/plugins/editor.lua
@@ -174,4 +174,22 @@ return {
       { "<leader>st", "<cmd>TodoTelescope<cr>",                            desc = "Todo" },
     },
   },
+  -- folding
+  {
+    "kevinhwang91/nvim-ufo",
+    dependencies = "kevinhwang91/promise-async",
+    config = function()
+      vim.keymap.set("n", "zR", require("ufo").openAllFolds)
+      vim.keymap.set("n", "zM", require("ufo").closeAllFolds)
+      vim.keymap.set("n", "zr", require("ufo").openFoldsExceptKinds)
+    end,
+  },
+  -- shading
+  {
+    "sunjon/shade.nvim",
+    config = function()
+      require("shade").setup()
+      require("shade").toggle()
+    end,
+  },
 }